+++ src/DTO/OrderData.php
@@ -126,6 +126,9 @@ class OrderData implements JsonSerializable
     /** @var ?string */
     private $title;
 
+    /** @var ?string */
+    private $expiresAt;
+
     public function __construct(
         Amount $amount,
         string $redirectUrl,
@@ -403,6 +406,18 @@ class OrderData implements JsonSerializable
         $this->title = $title;
     }
 
+    public function getExpiresAt(): ?string
+    {
+        return $this->expiresAt;
+    }
+
+    public function setExpiresAt(?string $expiresAt): self
+    {
+        $this->expiresAt = $expiresAt;
+
+        return $this;
+    }
+
     public function jsonSerialize(): array
     {
         $lines = [];
@@ -450,6 +465,7 @@ class OrderData implements JsonSerializable
             'lines' => $lines,
             'payment' => $this->getPayment()->jsonSerialize(),
             'consumerDateOfBirth' => $this->getConsumerDateOfBirth(),
+            'expiresAt' => $this->getExpiresAt(),
         ];
 
         if ($this->billingPhoneNumber) {